\name{RepDbLite.Hsapiens.2007}
\docType{package}

\alias{RepDbLite.Hsapiens.2007-package}
\alias{RepDbLite.Hsapiens.2007}
\alias{RepDbLite.Hsapiens.2007}


\title{RepBase-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at RepBase on
  Sat Oct 31 10:27:02 2015 and based on the RepBase20_07
}

\author{Nobody}



\examples{
## load the library
##library(RepDbLite.Hsapiens.2007)
## list the contents that are loaded into memory
ls('package:RepDbLite.Hsapiens.2007')
## show the db object that is loaded by calling it's name
RepDbLite.Hsapiens.2007

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
